+++ imap_processing/tests/ultra/unit/conftest.py
@@ -6,7 +6,13 @@ from pathlib import Path
 import pytest
 
 from imap_processing import decom
+from imap_processing.cdf.defaults import GlobalConstants
 from imap_processing.ultra.l0.decom_ultra import process_ultra_apids
+from imap_processing.ultra.l0.ultra_utils import (
+    ULTRA_AUX,
+    ULTRA_EVENTS,
+)
+from imap_processing.ultra.l1a import ultra_l1a
 from imap_processing.utils import group_by_apid
 
 
@@ -36,7 +42,7 @@ def ccsds_path_events():
     )
 
 
-@pytest.fixture()
+@pytest.fixture(scope="session")
 def ccsds_path_theta_0():
     """Returns the ccsds directory."""
     return (
@@ -63,7 +69,7 @@ def ccsds_path_tof():
     )
 
 
-@pytest.fixture()
+@pytest.fixture(scope="session")
 def xtce_path():
     """Returns the xtce image rates directory."""
     return (
@@ -178,3 +184,27 @@ def events_fsw_comparison_theta_0():
         / "l0"
         / filename
     )
+
+
+@pytest.fixture()
+def de_dataset(ccsds_path_theta_0, xtce_path):
+    """L1A test data"""
+    packets = decom.decom_packets(ccsds_path_theta_0, xtce_path)
+    grouped_data = group_by_apid(packets)
+    decom_ultra_events = process_ultra_apids(
+        grouped_data[ULTRA_EVENTS.apid[0]], ULTRA_EVENTS.apid[0]
+    )
+    decom_ultra_aux = process_ultra_apids(
+        grouped_data[ULTRA_AUX.apid[0]], ULTRA_AUX.apid[0]
+    )
+    dataset = ultra_l1a.create_dataset(
+        {
+            ULTRA_EVENTS.apid[0]: decom_ultra_events,
+            ULTRA_AUX.apid[0]: decom_ultra_aux,
+        }
+    )
+    # Remove start_type with fill values
+    l1a_de_dataset = dataset.where(
+        dataset["START_TYPE"] != GlobalConstants.INT_FILLVAL, drop=True
+    )
+    return l1a_de_dataset
