+++ imap_processing/tests/ultra/unit/test_ultra_l1b.py
@@ -2,17 +2,8 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from imap_processing import decom
-from imap_processing.cdf.defaults import GlobalConstants
-from imap_processing.ultra.l0.decom_ultra import process_ultra_apids
-from imap_processing.ultra.l0.ultra_utils import (
-    ULTRA_AUX,
-    ULTRA_EVENTS,
-)
-from imap_processing.ultra.l1a import ultra_l1a
 from imap_processing.ultra.l1b.ultra_l1b import ultra_l1b
 from imap_processing.ultra.utils.ultra_l1_utils import create_dataset
-from imap_processing.utils import group_by_apid
 
 
 @pytest.fixture()
@@ -75,30 +66,6 @@ def mock_data_l1b_dict():
     return data_dict
 
 
-@pytest.fixture()
-def de_dataset(ccsds_path_theta_0, xtce_path):
-    """L1A test data"""
-    packets = decom.decom_packets(ccsds_path_theta_0, xtce_path)
-    grouped_data = group_by_apid(packets)
-    decom_ultra_events = process_ultra_apids(
-        grouped_data[ULTRA_EVENTS.apid[0]], ULTRA_EVENTS.apid[0]
-    )
-    decom_ultra_aux = process_ultra_apids(
-        grouped_data[ULTRA_AUX.apid[0]], ULTRA_AUX.apid[0]
-    )
-    dataset = ultra_l1a.create_dataset(
-        {
-            ULTRA_EVENTS.apid[0]: decom_ultra_events,
-            ULTRA_AUX.apid[0]: decom_ultra_aux,
-        }
-    )
-    # Remove start_type with fill values
-    l1a_de_dataset = dataset.where(
-        dataset["START_TYPE"] != GlobalConstants.INT_FILLVAL, drop=True
-    )
-    return l1a_de_dataset
-
-
 def test_create_dataset(mock_data_l1b_dict):
     """Tests that dataset is created as expected."""
     dataset = create_dataset(mock_data_l1b_dict, "imap_ultra_l1b_45sensor-de", "l1b")
@@ -106,7 +73,7 @@ def test_create_dataset(mock_data_l1b_dict):
     assert "epoch" in dataset.coords
     assert dataset.coords["epoch"].dtype == "datetime64[ns]"
     assert dataset.attrs["Logical_source"] == "imap_ultra_l1b_45sensor-de"
-    assert dataset["x_front"].attrs["UNITS"] == "hundredths of a millimeter"
+    assert dataset["x_front"].attrs["UNITS"] == "mm / 100"
     np.testing.assert_array_equal(dataset["x_front"], np.zeros(3))
